refactor(aws): Rename head-objects to head
