+++ packages/aws/src/lib/storage/head.ts
